+++ src/components/EditSections/EditUserInfo/EditUserInfo.js
@@ -303,8 +303,7 @@ class EditUserInfo extends React.Component {
                 dataOptions={patronGroupOptions}
                 defaultValue={initialValues.patronGroup}
                 aria-required="true"
-                required
-                disabled={disabled}
+                required={!disabled}
               />
               <OnChange name="patronGroup">
                 {(selectedPatronGroup) => {