+++ src/views/UserEdit/UserEdit.js
@@ -215,7 +215,7 @@ class UserEdit extends React.Component {
     }
 
     user.personal.addresses = toUserAddresses(user.personal.addresses); // eslint-disable-line no-param-reassign
-    user.personal.email = user.personal.email.trim();
+    user.personal.email = user.personal.email?.trim();
     user.departments = compact(user.departments);
 
     const { proxies, sponsors, permissions, servicePoints, preferredServicePoint } = user;